+++ source/Tubeshade.Server/Configuration/Startup/Migrations/ChannelImagesMigration.cs
@@ -4,6 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Npgsql;
 using Tubeshade.Data;
+using Tubeshade.Data.AccessControl;
 using Tubeshade.Data.Identity;
 using Tubeshade.Data.Media;
 using Tubeshade.Data.Tasks;
@@ -11,22 +12,25 @@ using static System.Data.IsolationLevel;
 
 namespace Tubeshade.Server.Configuration.Startup.Migrations;
 
-public sealed class TrackFileMigration : IApplicationMigration
+internal sealed class ChannelImagesMigration : IApplicationMigration
 {
     private readonly NpgsqlConnection _connection;
     private readonly UserRepository _userRepository;
-    private readonly LibraryRepository _libraryRepository;
+    private readonly ChannelRepository _channelRepository;
+    private readonly ImageFileRepository _imageRepository;
     private readonly TaskRepository _taskRepository;
 
-    public TrackFileMigration(
+    public ChannelImagesMigration(
         NpgsqlConnection connection,
         UserRepository userRepository,
-        LibraryRepository libraryRepository,
+        ChannelRepository channelRepository,
+        ImageFileRepository imageRepository,
         TaskRepository taskRepository)
     {
         _connection = connection;
         _userRepository = userRepository;
-        _libraryRepository = libraryRepository;
+        _channelRepository = channelRepository;
+        _imageRepository = imageRepository;
         _taskRepository = taskRepository;
     }
 
@@ -34,25 +38,33 @@ public sealed class TrackFileMigration : IApplicationMigration
     public async ValueTask MigrateAsync(CancellationToken cancellationToken)
     {
         Guid userId;
-        List<LibraryEntity> libraries;
+        List<ChannelEntity> channels;
 
         await using (var transaction = await _connection.OpenAndBeginTransaction(ReadCommitted, cancellationToken))
         {
             userId = await _userRepository.GetSystemUserId(transaction);
-            libraries = await _libraryRepository.GetAsync(userId, transaction);
+            channels = await _channelRepository.GetAsync(userId, transaction);
 
             await transaction.CommitAsync(cancellationToken);
         }
 
-        foreach (var library in libraries)
+        foreach (var channel in channels)
         {
             await using var transaction = await _connection.OpenAndBeginTransaction(cancellationToken);
+            var images = await _imageRepository.GetForChannel(channel.Id, userId, Access.Read, transaction, cancellationToken);
+            if (images is not [])
+            {
+                continue;
+            }
+
+            var libraryId = await _channelRepository.GetPrimaryLibraryId(channel.Id, transaction);
             var tasks = await _taskRepository.GetRunningTasks(
                 new TaskParameters
                 {
                     UserId = userId,
-                    LibraryId = library.Id,
-                    Type = TaskType.RefreshTrackFiles,
+                    LibraryId = libraryId,
+                    Type = TaskType.Index,
+                    Url = channel.ExternalUrl,
                     Limit = 1,
                     Offset = 0,
                 },
@@ -61,7 +73,7 @@ public sealed class TrackFileMigration : IApplicationMigration
 
             if (tasks is [])
             {
-                var task = TaskEntity.RefreshTracks(library.Id, userId);
+                var task = TaskEntity.Index(libraryId, userId, channel.Id, null, channel.ExternalUrl);
                 if (await _taskRepository.TryAddTask(task, transaction) is { } taskId)
                 {
                     await _taskRepository.TriggerTask(taskId, TaskSource.Schedule, transaction);